[video_player] Add v2 embedding support (#2226)
- Adds support in the plugin
- Adds a v2 embedding to the example app
- Fixes a broken remote example in the example app
- Increments the Flutter SDK dependency
- Increments the version
- Adds e2e tests for some simple use cases of the plugin
