+++ packages/video_player/example/pubspec.yaml
@@ -8,7 +8,9 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
+  flutter_driver:
+    sdk: flutter
+  e2e: "^0.2.0"
   video_player:
     path: ../
